[TASK] (vendor_Seitosh): Seitosh vendor drop from 1.2.2013
Replay vendor drop as took place in the original subversion repository of
DENISE. Source code from the subversion repository of TFSoftware was copied to
vendor branches aff, cseife, fourier, and stfinv.

aff:     subversion revision 372
cseife:  subversion revision 369
fourier: subversion revision 375
stfinv:  subversion revision 378

Source code is copied from tag directories 20130201

Notice: The contents of this commit are not in sync with the corresponding
export tags as present in the Seitosh repository. This must be due to
unintentional modification of files when copying contents as tagged for export
in TFSoftware to the vendor branch in the subversion repository of (formerly
called FWI_elastic/DENISE).