[FIX] (import_Seitosh): set variables in vendor Makefiles
The mechanism in contrib/Makefile_var which was used to set the make variables
expected by the Seitosh vendor code only worked for Makefiles in the next
subdirectory level below contrib. This was due to relative paths used in the
definitions. By passing the relative path of the contrib directory to
Makefile_var in a make variable, allows Makefile_var to the path values
appropriately for any Makefile in any directory, which include Makefile_var