[TASK] (import_Seitosh): read CONTRIBDIR from automatic variable
Since vendor Makefiles make use of special GNU make features, we can safely
rely on GNU make being in operation. GNU make provides the variable
MAKEFILE_LIST from which the relative path can be extracted at which
Makefile_var was found. This exactly is the path of the contrib directory,
such that the user (calling Makefile) needs not set this variable.

Currently the calling Makefiles still set a variable CONTRIBDIR which then
simply is overwritten in Makefile_var. Such the setting of CONTRIBDIR is
unnecessarily doubled.