[TASK] (import_Seitosh): discard contrib/aff/tests
Code in contrib/aff/tests is not required as part of Seitosh.
However, the test code requires additional settings like proper interlanguage
linking (C++ and Fortran). To reduce the constraints on the installation
environment, discard the subdirectory tests in contrib/aff entirely, as well
as all references to the test code made in the Makefile.