Fix conjugate gradient and low pass filtering
This fix will reset the conjugate gradient history each time a frequency or workflow stage is switched.
If such a stage is switched we will have a "new" objective function, thus, it makes no sense to use the old gradients for speeding up the convergence.
This bug was the reason why the objective function could not be reduced for 2-3 iterations after a new frequency stage was appplied.